fix: add blockexplorer url for pool page actions (#239)
