+++ osv-scanner.toml
@@ -14,3 +14,40 @@ reason = "PPR (Partial Prerendering) is not used in this project — the vulnera
 [[IgnoredVulns]]
 id = "GHSA-848j-6mx2-7j84"
 reason = "No patch available — transitive dep of ethers v5 (@ethersproject/signing-key). Low practical risk for client-side usage"
+
+# GHSA-2g4f-4pwh-qvx6: ajv ReDoS when using `$data` option
+# ajv@6.12.6 is the latest on the v6 line. Pinned by eslint which requires ^6.12.4.
+# Dev-only dependency (linting toolchain). The ReDoS requires the `$data` option which
+# eslint does not use.
+[[IgnoredVulns]]
+id = "GHSA-2g4f-4pwh-qvx6"
+reason = "ajv@6.12.6 pinned by eslint — latest on v6 line, no fix available. Dev-only; eslint does not use the vulnerable $data option"
+
+# GHSA-378v-28hj-76wf: bn.js infinite loop
+# bn.js@4.12.2 is pinned by elliptic@^4.11.9, a transitive dep of ethers v5.
+# Cannot override to v5 without breaking elliptic's crypto operations.
+# The infinite loop requires crafted input to bn.js arithmetic — not reachable from
+# user input in this application.
+[[IgnoredVulns]]
+id = "GHSA-378v-28hj-76wf"
+reason = "bn.js@4.12.2 pinned by elliptic (ethers v5) — cannot bump major without breakage. Vulnerable code path not reachable from user input"
+
+# GHSA-23c5-xmqv-rm74: minimatch ReDoS via nested *() extglobs
+# GHSA-3ppc-4f35-3m26: minimatch ReDoS via repeated wildcards
+# GHSA-7r86-cg39-jmmj: minimatch ReDoS via multiple GLOBSTAR segments
+# minimatch@3.1.2 is the latest on v3 — pinned by eslint (dev-only).
+# minimatch@9.0.0 is pinned by @turbo/gen (dev-only).
+# minimatch@9.0.5 is the latest on v9 — pinned by @sentry/node and typescript-eslint.
+# All ReDoS vulns require malicious glob patterns; in these contexts patterns come from
+# config files, not user input.
+[[IgnoredVulns]]
+id = "GHSA-23c5-xmqv-rm74"
+reason = "minimatch v3/v9 pinned by eslint, @turbo/gen, @sentry/node — no fix on these major lines. Glob patterns are config-sourced, not user input"
+
+[[IgnoredVulns]]
+id = "GHSA-3ppc-4f35-3m26"
+reason = "minimatch v3/v9 pinned by eslint, @turbo/gen, @sentry/node — no fix on these major lines. Glob patterns are config-sourced, not user input"
+
+[[IgnoredVulns]]
+id = "GHSA-7r86-cg39-jmmj"
+reason = "minimatch v3/v9 pinned by eslint, @turbo/gen, @sentry/node — no fix on these major lines. Glob patterns are config-sourced, not user input"
