+++ package.json
@@ -68,20 +68,25 @@
       "@metamask/sdk@>=0.16.0 <=0.33.0": ">=0.33.1",
       "@metamask/sdk-communication-layer@>=0.16.0 <=0.33.0": ">=0.33.1",
       "@tanstack/react-query": "^5.83.0",
+      "ajv@>=8.0.0 <8.18.0": ">=8.18.0",
       "axios@>=1.13.0 <1.13.5": ">=1.13.5",
+      "basic-ftp@<5.2.0": ">=5.2.0",
+      "bn.js@>=5.0.0 <5.2.3": ">=5.2.3",
       "color-string@<1.5.5": ">=1.5.5",
       "diff": ">=5.2.0",
       "form-data@<2.5.4": ">=2.5.4",
       "glob@<11.1.0": ">=11.1.0",
       "h3@<1.15.5": ">=1.15.5",
-      "hono@<4.11.7": ">=4.11.7",
+      "hono@<4.12.3": ">=4.12.3",
       "js-yaml@<4.1.1": ">=4.1.1",
       "linkifyjs": "^4.3.2",
       "lodash": "^4.17.23",
       "markdown-it@<14.1.1": ">=14.1.1",
       "mdast-util-to-hast": "^13.2.1",
+      "minimatch@>=10.0.0 <10.2.4": ">=10.2.4",
       "next": "^15.5.10",
       "preact": ">=10.28.2",
+      "rollup@>=4.0.0 <4.59.0": ">=4.59.0",
       "tmp": ">=0.2.4",
       "wagmi": "^2.16.9",
       "zod": "^4.1.13"
